fix: re-write migration to fix error introduced by return to built-in validators
A custom validator was introduced as part of the migration to a custom user
model. The custom validator was removed because it was offered in contrib.auth.
See: 382b0f57e77b619561ab390b1e10dc0afaa6b761. This introduced an
inconsistency in initial migration; the inconsistency is corrected here.